ImageManipV2 RAW16 support

+++ src/pipeline/datatype/ImageManipConfigV2.cpp
@@ -73,11 +73,11 @@ ImageManipConfigV2& ImageManipConfigV2::setColormap(Colormap colormap) {
     base.setColormap(colormap);
     return *this;
 }
-ImageManipConfigV2& ImageManipConfigV2::setBackgroundColor(uint8_t red, uint8_t green, uint8_t blue) {
+ImageManipConfigV2& ImageManipConfigV2::setBackgroundColor(uint32_t red, uint32_t green, uint32_t blue) {
     base.setBackgroundColor(red, green, blue);
     return *this;
 }
-ImageManipConfigV2& ImageManipConfigV2::setBackgroundColor(uint8_t val) {
+ImageManipConfigV2& ImageManipConfigV2::setBackgroundColor(uint32_t val) {
     base.setBackgroundColor(val);
     return *this;
 }
